// admin/src/app/services/api/models/CreatePersonDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreatePersonDto = {
  id?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  emailAddress?: string | null;
  mobileNumber?: string | null;
  dateOfBirth?: Date | null;
  addressId?: string | null;
  streeAddress?: string | null;
  city?: string | null;
  postalCode?: string | null;
  country?: string | null;
};